% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{myfunction}
\alias{myfunction}
\title{This is new one I created}
\usage{
myfunction(a, b)
}
\arguments{
\item{a}{random number}

\item{b}{random number 2}
}
\value{
c this is the one we will have it
}
\description{
This is new one I created
}
\examples{
myfunction(a=5,b=7)
}
