% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deneme.R
\name{myfunction1}
\alias{myfunction1}
\title{I am trying to learn the Roxygen}
\usage{
myfunction1(a, b)
}
\arguments{
\item{a}{random number}

\item{b}{random number}
}
\value{

}
\description{
I am trying to learn the Roxygen
}
